Separate ALB alarms module and ALB Target Group alarms module
- Create unit-tests for ALB alarms and ALB Taget Group alarms
