+++ core/alarms.js
@@ -13,6 +13,7 @@ const ecsAlarms = require('./alarms-ecs')
 const snsAlarms = require('./alarms-sns')
 const ruleAlarms = require('./alarms-eventbridge')
 const albAlarms = require('./alarms-alb')
+const albTargetAlarms = require('./alarms-alb-target-group')
 
 module.exports = function alarms (alarmConfig, functionAlarmConfigs, context) {
   const {
@@ -25,7 +26,8 @@ module.exports = function alarms (alarmConfig, functionAlarmConfigs, context) {
     ECS: ecsConfig,
     SNS: snsConfig,
     Events: ruleConfig,
-    ApplicationELB: albConfig
+    ApplicationELB: albConfig,
+    ApplicationELBTarget: albTargetConfig
   } = cascade(alarmConfig)
 
   const cascadedFunctionAlarmConfigs = applyAlarmConfig(lambdaConfig, functionAlarmConfigs)
@@ -38,8 +40,9 @@ module.exports = function alarms (alarmConfig, functionAlarmConfigs, context) {
   const { createECSAlarms } = ecsAlarms(ecsConfig, context)
   const { createSNSAlarms } = snsAlarms(snsConfig, context)
   const { createRuleAlarms } = ruleAlarms(ruleConfig, context)
+  const { createALBAlarms } = albAlarms(albConfig, context)
+  const { createALBTargetAlarms } = albTargetAlarms(albTargetConfig, context)
 
-  const { createALBAlarms } = albAlarms(albConfig, context, serverless)
   return {
     addAlarms
   }
@@ -62,6 +65,7 @@ module.exports = function alarms (alarmConfig, functionAlarmConfigs, context) {
       snsConfig.enabled && createSNSAlarms(cfTemplate)
       ruleConfig.enabled && createRuleAlarms(cfTemplate)
       albConfig.enabled && createALBAlarms(cfTemplate)
+      albTargetConfig.enabled && createALBTargetAlarms(cfTemplate)
     }
   }
 }
