+++ core/config-schema.js
@@ -23,7 +23,6 @@ const supportedAlarms = {
   Events: ['FailedInvocations', 'ThrottledRules'],
   ApplicationELB: ['HTTPCode_ELB_5XX_Count', 'RejectedConnectionCount'],
   ApplicationELBTarget: ['HTTPCode_Target_5XX_Count', 'UnHealthyHostCount', 'LambdaInternalError', 'LambdaUserError']
-
 }
 
 const supportedWidgets = {
