+++ serverless-plugin/alarms-alb.js
@@ -1,20 +1,6 @@
 'use strict'
 
 const { makeResourceName, getStatisticName } = require('./util')
-/**
-   * Given CloudFormation syntax for an Target Group, derive CloudFormation syntax for
-   * the LoadBalancer LogicalId name
-   *
-   * @param  loadBalancerResources  syntax for an Load Balancer Application
-   * @returns CloudFormation syntax for the Load Balancer Resources
-   */
-function resolveLoadBalancerLogicalIdName (loadBalancerResources) {
-  for (const key in loadBalancerResources) {
-    if (loadBalancerResources[key].Type === 'AWS::ElasticLoadBalancingV2::LoadBalancer') {
-      return key
-    }
-  }
-}
 
 /**
  * @param {object} albAlarmConfig The fully resolved alarm configuration
@@ -54,49 +40,6 @@ module.exports = function ALBlarms (albAlarmConfig, context) {
         cfTemplate.addResource(rejectedConnectionCount.resourceName, rejectedConnectionCount.resource)
       }
     }
-
-    const targetGroupResources = cfTemplate.getResourcesByType(
-      'AWS::ElasticLoadBalancingV2::TargetGroup'
-    )
-    for (const [targetGroupResourceName, targetGroupResource] of Object.entries(targetGroupResources)) {
-      const loadBalancerName = resolveLoadBalancerLogicalIdName(loadBalancerResources)
-      if (albAlarmConfig.HTTPCode_Target_5XX_Count && albAlarmConfig.HTTPCode_Target_5XX_Count.enabled) {
-        const httpCodeTarget5XXCount = createHTTPCodeTarget5XXCountAlarm(
-          targetGroupResourceName,
-          targetGroupResource,
-          loadBalancerName,
-          albAlarmConfig.HTTPCode_Target_5XX_Count
-        )
-        cfTemplate.addResource(httpCodeTarget5XXCount.resourceName, httpCodeTarget5XXCount.resource)
-      }
-      if (albAlarmConfig.UnHealthyHostCount && albAlarmConfig.UnHealthyHostCount.enabled) {
-        const unHealthyHostCount = createUnHealthyHostCountAlarm(
-          targetGroupResourceName,
-          targetGroupResource,
-          loadBalancerName,
-          albAlarmConfig.UnHealthyHostCount
-        )
-        cfTemplate.addResource(unHealthyHostCount.resourceName, unHealthyHostCount.resource)
-      }
-      if (albAlarmConfig.LambdaInternalError && albAlarmConfig.LambdaInternalError.enabled) {
-        const lambdaInternalError = createLambdaInternalErrorAlarm(
-          targetGroupResourceName,
-          targetGroupResource,
-          loadBalancerName,
-          albAlarmConfig.LambdaInternalError
-        )
-        cfTemplate.addResource(lambdaInternalError.resourceName, lambdaInternalError.resource)
-      }
-      if (albAlarmConfig.LambdaUserError && albAlarmConfig.LambdaUserError.enabled) {
-        const lambdaUserError = createLambdaUserErrorAlarm(
-          targetGroupResourceName,
-          targetGroupResource,
-          loadBalancerName,
-          albAlarmConfig.LambdaUserError
-        )
-        cfTemplate.addResource(lambdaUserError.resourceName, lambdaUserError.resource)
-      }
-    }
   }
 
   function createLoadBalancerAlarm (
@@ -113,53 +56,8 @@ module.exports = function ALBlarms (albAlarmConfig, context) {
     treatMissingData
   ) {
     const loadBalancerFullName = { 'Fn::GetAtt': [loadBalancerResourceName, 'LoadBalancerFullName'] }
-    console.log('1', loadBalancerFullName)
     const metricProperties = {
       Dimensions: [{ Name: 'LoadBalancer', Value: loadBalancerFullName }],
-      MetricName: metricName,
-      Namespace: 'AWS/ApplicationELB',
-      Period: period,
-      Statistic: statistic,
-      ExtendedStatistic: extendedStatistic
-    }
-
-    return {
-      Type: 'AWS::CloudWatch::Alarm',
-      Properties: {
-        ActionsEnabled: true,
-        AlarmActions: context.alarmActions,
-        AlarmName: alarmName,
-        AlarmDescription: alarmDescription,
-        EvaluationPeriods: evaluationPeriods,
-        ComparisonOperator: comparisonOperator,
-        Threshold: threshold,
-        TreatMissingData: treatMissingData,
-        ...metricProperties
-      }
-    }
-  }
-
-  function createLoadBalancerTargetAlarm (
-    alarmName,
-    alarmDescription,
-    targetGroupResourceName, // Logical ID of the CloudFormation Target Group Resource
-    loadBalancerName,
-    comparisonOperator,
-    threshold,
-    metricName,
-    statistic,
-    period,
-    extendedStatistic,
-    evaluationPeriods,
-    treatMissingData
-  ) {
-    const targetGroupFullName = { 'Fn::GetAtt': [targetGroupResourceName, 'TargetGroupFullName'] }
-    console.log('3', targetGroupFullName)
-    console.log(targetGroupResourceName)
-    const loadBalancerFullName = { 'Fn::GetAtt': [loadBalancerName, 'LoadBalancerFullName'] }
-    console.log('2', loadBalancerFullName)
-    const metricProperties = {
-      Dimensions: [{ Name: 'TargetGroup', Value: targetGroupFullName }, { Name: 'LoadBalancer', Value: loadBalancerFullName }],
       MetricName: metricName,
       Namespace: 'AWS/ApplicationELB',
       Period: period,
@@ -214,90 +112,6 @@ module.exports = function ALBlarms (albAlarmConfig, context) {
         config.ComparisonOperator,
         threshold,
         'RejectedConnectionCount', // metricName
-        config.Statistic,
-        config.Period,
-        config.ExtendedStatistic,
-        config.EvaluationPeriods,
-        config.TreatMissingData
-      )
-    }
-  }
-
-  function createHTTPCodeTarget5XXCountAlarm (targetGroupResourceName, targetGroupResource, loadBalancerName, config) {
-    const threshold = config.Threshold
-    return {
-      resourceName: makeResourceName('LoadBalancer', targetGroupResourceName, 'HTTPCodeTarget5XXCount'),
-      resource: createLoadBalancerTargetAlarm(
-        `LoadBalancerHTTPCodeTarget5XXCountAlarm_${targetGroupResourceName}`,
-        `LoadBalancer HTTP Code Target 5XX Count ${getStatisticName(config)} for ${targetGroupResourceName} breaches ${threshold}`,
-        targetGroupResourceName,
-        loadBalancerName,
-        config.ComparisonOperator,
-        threshold,
-        'HTTPCode_Target_5XX_Count',
-        config.Statistic,
-        config.Period,
-        config.ExtendedStatistic,
-        config.EvaluationPeriods,
-        config.TreatMissingData
-      )
-    }
-  }
-
-  function createUnHealthyHostCountAlarm (targetGroupResourceName, targetGroupResource, loadBalancerName, config) {
-    const threshold = config.Threshold
-    return {
-      resourceName: makeResourceName('LoadBalancer', targetGroupResourceName, 'UnHealthyHostCount'),
-      resource: createLoadBalancerTargetAlarm(
-        `LoadBalancerUnHealthyHostCountAlarm_${targetGroupResourceName}`,
-        `LoadBalancer UnHealthy Host Count ${getStatisticName(config)} for ${targetGroupResourceName} breaches ${threshold}`,
-        targetGroupResourceName,
-        loadBalancerName,
-        config.ComparisonOperator,
-        threshold,
-        'UnHealthyHostCount',
-        config.Statistic,
-        config.Period,
-        config.ExtendedStatistic,
-        config.EvaluationPeriods,
-        config.TreatMissingData
-      )
-    }
-  }
-
-  function createLambdaInternalErrorAlarm (targetGroupResourceName, targetGroupResource, loadBalancerName, config) {
-    const threshold = config.Threshold
-    return {
-      resourceName: makeResourceName('LoadBalancer', targetGroupResourceName, 'LambdaInternalError'),
-      resource: createLoadBalancerTargetAlarm(
-        `LoadBalancerLambdaInternalErrorAlarm_${targetGroupResourceName}`,
-        `LoadBalancer Lambda Internal Error ${getStatisticName(config)} for ${targetGroupResourceName} breaches ${threshold}`,
-        targetGroupResourceName,
-        loadBalancerName,
-        config.ComparisonOperator,
-        threshold,
-        'LambdaInternalError',
-        config.Statistic,
-        config.Period,
-        config.ExtendedStatistic,
-        config.EvaluationPeriods,
-        config.TreatMissingData
-      )
-    }
-  }
-
-  function createLambdaUserErrorAlarm (targetGroupResourceName, targetGroupResource, loadBalancerName, config) {
-    const threshold = config.Threshold
-    return {
-      resourceName: makeResourceName('LoadBalancer', targetGroupResourceName, 'LambdaUserError'),
-      resource: createLoadBalancerTargetAlarm(
-        `LoadBalancerLambdaUserErrorAlarm_${targetGroupResourceName}`,
-        `LoadBalancer Lambda User Error ${getStatisticName(config)} for ${targetGroupResourceName} breaches ${threshold}`,
-        targetGroupResourceName,
-        loadBalancerName,
-        config.ComparisonOperator,
-        threshold,
-        'LambdaUserError',
         config.Statistic,
         config.Period,
         config.ExtendedStatistic,
